Auto stash before merge of "master" and "origin/master"
