+++ cs/TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -67,7 +67,6 @@
     <Compile Include="Rectangle_should.cs" />
     <Compile Include="Size.cs" />
     <Compile Include="TagsCloudVisualization_should.cs" />
-    <Compile Include="Visualizer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
